requirements.txt
pyyaml
opencv-python
numpy
matplotlib